Cleanup and Prepare for Testing
