+++ tests/e2e.spec.ts
@@ -52,7 +52,7 @@ describe("Near Safe Requests", () => {
     await expect(adapter.policyForChainId(100)).resolves.not.toThrow();
   });
 
-  it("adapter: encodeEvmTx", async () => {
+  it.only("adapter: encodeEvmTx", async () => {
     await expect(
       adapter.encodeSignRequest({
         method: "eth_sendTransaction",
@@ -67,44 +67,21 @@ describe("Near Safe Requests", () => {
         ],
       })
     ).resolves.not.toThrow();
-  });
 
-  it("adapter: decodeTxData", async () => {
-    // setup:
-    const chainId = 11155111;
-    const expectedMessage =
-      "Welcome to OpenSea!\n" +
-      "\n" +
-      "Click to sign in and accept the OpenSea Terms of Service (https://opensea.io/tos) and Privacy Policy (https://opensea.io/privacy).\n" +
-      "\n" +
-      "This request will not trigger a blockchain transaction or cost any gas fees.\n" +
-      "\n" +
-      "Wallet address:\n" +
-      "0xf057e37024abe7e6bc04fb4f00978613b5ca0241\n" +
-      "\n" +
-      "Nonce:\n" +
-      "aca09a1c-a800-4d71-98ed-547f7c59370c";
-
-    const signRequest = await adapter.encodeSignRequest({
-      method: "personal_sign",
-      chainId,
-      params: [
-        "0x57656c636f6d6520746f204f70656e536561210a0a436c69636b20746f207369676e20696e20616e642061636365707420746865204f70656e536561205465726d73206f662053657276696365202868747470733a2f2f6f70656e7365612e696f2f746f732920616e64205072697661637920506f6c696379202868747470733a2f2f6f70656e7365612e696f2f70726976616379292e0a0a5468697320726571756573742077696c6c206e6f742074726967676572206120626c6f636b636861696e207472616e73616374696f6e206f7220636f737420616e792067617320666565732e0a0a57616c6c657420616464726573733a0a3078663035376533373032346162653765366263303466623466303039373836313362356361303234310a0a4e6f6e63653a0a61636130396131632d613830302d346437312d393865642d353437663763353933373063",
-        "0xf057e37024abe7e6bc04fb4f00978613b5ca0241",
-      ],
-    });
-
-    expect(signRequest.evmData).toStrictEqual({
-      chainId,
-      hash: "0xb3a14f9bd21518d7da23dba01ddf7c7ef45795ca1515f1b41b6f3455c862e22d",
-      data: expectedMessage,
-    });
-
-    expect(adapter.decodeTxData(signRequest.evmData)).toStrictEqual({
+    const typedDataString = `{
+        \"types\":{\"SafeTx\":[{\"name\":\"to\",\"type\":\"address\"},{\"name\":\"value\",\"type\":\"uint256\"},{\"name\":\"data\",\"type\":\"bytes\"},{\"name\":\"operation\",\"type\":\"uint8\"},{\"name\":\"safeTxGas\",\"type\":\"uint256\"},{\"name\":\"baseGas\",\"type\":\"uint256\"},{\"name\":\"gasPrice\",\"type\":\"uint256\"},{\"name\":\"gasToken\",\"type\":\"address\"},{\"name\":\"refundReceiver\",\"type\":\"address\"},{\"name\":\"nonce\",\"type\":\"uint256\"}],\"EIP712Domain\":[{\"name\":\"chainId\",\"type\":\"uint256\"},{\"name\":\"verifyingContract\",\"type\":\"address\"}]},\"domain\":{\"chainId\":\"0xaa36a7\",\"verifyingContract\":\"0x7fa8e8264985c7525fc50f98ac1a9b3765405489\"},\"primaryType\":\"SafeTx\",\"message\":{\"to\":\"0x102543f7e6b5786a444cc89ff73012825d13000d\",\"value\":\"100000000000000000\",\"data\":\"0x\",\"operation\":\"0\",\"safeTxGas\":\"0\",\"baseGas\":\"0\",\"gasPrice\":\"0\",\"gasToken\":\"0x0000000000000000000000000000000000000000\",\"refundReceiver\":\"0x0000000000000000000000000000000000000000\",\"nonce\":\"0\"}
+        }`;
+    const { evmData } = await adapter.encodeSignRequest({
       chainId: 11155111,
-      costEstimate: "0",
-      transactions: [],
-      message: expectedMessage,
+      method: "eth_signTypedData_v4",
+      params: ["0x102543f7e6b5786a444cc89ff73012825d13000d", typedDataString],
+    });
+
+    expect(evmData).toStrictEqual({
+      chainId: 11155111,
+      // This will also be fixed by: https://github.com/BitteProtocol/near-ca/pull/142
+      hash: "0x",
+      data: typedDataString,
     });
   });
 });
